Modified   SourceCode/Engine/Binaries/Alllocators.data Modified   SourceCode/Engine/Engine/MemoryManagement/Allocator/AllocatorBase.h Modified   SourceCode/Engine/Engine/MemoryManagement/Allocator/CustomAllocator.cpp Modified   SourceCode/Engine/Engine/MemoryManagement/Allocator/CustomAllocator.h Modified   SourceCode/Engine/Engine/MemoryManagement/Allocator/DefaultAllocator.cpp Modified   SourceCode/Engine/Engine/MemoryManagement/Allocator/DefaultAllocator.h Modified   SourceCode/Engine/Engine/MemoryManagement/Allocator/FrameAllocator.cpp Modified   SourceCode/Engine/Engine/MemoryManagement/Allocator/Initializer.cpp Modified   SourceCode/Engine/Engine/MemoryManagement/Allocator/Initializer.h Modified   SourceCode/Engine/Engine/Test/TestLauncher/TestLauncher.cpp

+++ SourceCode/Engine/Engine/MemoryManagement/Allocator/AllocatorBase.h
@@ -36,7 +36,7 @@ namespace Engine
 				virtual void Deallocate(byte* Address) = 0;
 				virtual bool TryDeallocate(byte* Address) = 0;
 
-				virtual uint32 GetReservedSize(void) const
+				virtual uint64 GetReservedSize(void) const
 				{
 					return 0;
 				}
